allow custom prefix and full username
Since the username length is limited to 32 characters, and many people log in using email with Cognito, and email lengths tend to be around 29 characters, I allow the use of a custom prefix and username. TODO: Put a disclaimer about the email length somewhere, warning people of this issue. I'm sad right now; none of this refactoring is needed If the OS allowed 64 characters in usernames.